plaid

+++ Api/Spendly.Mobile.BusinessLayer/Services/Finance/UserMerchantService.cs
@@ -13,10 +13,10 @@ using Spendly.Shared.Entities.UserManagement;
 using Spendly.Shared.Localization;
 using Spendly.Shared.ViewModels;
 
-public class MerchantService(
+public class UserMerchantService(
     IRepository<UserMerchant> userMerchantRepository,
     IRepository<Merchant> merchantRepository,
-    IRepository<UserCategory> categoryRepository,
+    IRepository<UserCategory> userCategoryRepository,
     RequestContextViewModel requestContextViewModel)
 {
     //TODO review
@@ -24,18 +24,12 @@ public class MerchantService(
     {
         var userMerchants = (await userMerchantRepository.ListAsync(x => x.UserId == requestContextViewModel.UserId.ToObjectId())).ToList();
 
-        if (request.IsUncategorized == true)
-        {
-            userMerchants = userMerchants.Where(x => x.UserCategoryId == null).ToList();
-        }
-
         if (userMerchants.Count == 0)
         {
             return FunctionResponse.Success(new List<MerchantListItemViewModel>());
         }
 
-        var categoryLookup = (await categoryRepository.ListAsync(x => x.UserId == requestContextViewModel.UserId.ToObjectId()))
-            .ToDictionary(x => x.Id, x => x.Name);
+        var allUserCategories = await userCategoryRepository.ListDictionaryAsync(userMerchants.Where(p=>p.UserCategoryId.HasValue).Select(p => p.UserCategoryId!.Value));
 
         var allMerchants = await merchantRepository.ListDictionaryAsync(userMerchants.Select(p => p.MerchantId));
 
@@ -45,8 +39,11 @@ public class MerchantService(
             Name = allMerchants[x.MerchantId].Name,
             Nickname = x.Nickname,
             CategoryId = x.UserCategoryId?.ToString(),
-            CategoryName = x.UserCategoryId != null && categoryLookup.TryGetValue(x.UserCategoryId.Value, out var name) ? name : "Uncategorized",
-        }).ToList();
+            CategoryName = x.UserCategoryId != null? allUserCategories[x.UserCategoryId.Value].Name : "",
+            IsOther = x.IsOther
+        })
+        .OrderBy(p => p.IsOther)
+            .ThenBy(p=>p.Name).ToList();
 
         return FunctionResponse.Success(items);
     }
@@ -61,26 +58,25 @@ public class MerchantService(
 
         var merchant = await userMerchantRepository.GetRequiredAsync(x => x.Id == merchantId && x.UserId == requestContextViewModel.UserId.ToObjectId());
 
+        if (merchant.IsOther)
+        {
+            return FunctionResponse.Failure(MessageCodes.CannotModifyOtherMerchant);
+        }
+
         await userMerchantRepository.DeleteAsync(merchant.Id);
         return FunctionResponse.Success();
     }
 
 
-    public async Task<FunctionResponse<MerchantDetailViewModel>> UpdateAsync(string id, MerchantUpdateRequestViewModel request)
+    public async Task<FunctionResponse<MerchantDetailViewModel>> UpdateAsync(string id, UserMerchantUpdateRequestViewModel request)
     {
-        var merchantId = id.ToObjectIdOrNull();
-        if (merchantId == null)
-        {
-            return FunctionResponse.Failure<MerchantDetailViewModel>(MessageCodes.MerchantNotFound);
-        }
-
-        var userMerchant = await userMerchantRepository.GetRequiredAsync(x => x.Id == merchantId && x.UserId == requestContextViewModel.UserId.ToObjectId());
+        var userMerchant = await userMerchantRepository.GetRequiredAsync(x => x.Id == id.ToObjectId() && x.UserId == requestContextViewModel.UserId.ToObjectId());
 
         userMerchant.UserCategoryId = request.CategoryId?.ToObjectId();
         userMerchant.Nickname = request.Nickname;
         await userMerchantRepository.UpdateAsync(userMerchant);
         
-        var merchant = await merchantRepository.GetRequiredAsync(x => x.Id == merchantId);
+        var merchant = await merchantRepository.GetRequiredAsync(x => x.Id == userMerchant.MerchantId);
 
         var response = new MerchantDetailViewModel
         {
@@ -88,6 +84,7 @@ public class MerchantService(
             Name = merchant.Name,
             Nickname = userMerchant.Nickname,
             CategoryId = userMerchant.UserCategoryId?.ToString(),
+            IsOther = userMerchant.IsOther,
         };
 
         return FunctionResponse.Success(response);
